backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==1.4.53
aiosqlite==0.19.0
redis==5.0.1
openai==1.12.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
numpy==1.24.3
scikit-learn==1.3.2
sentence-transformers==2.2.2